+++ src/view/vendingMachineView.js
